chore: move function to locationUtils
