+++ src/utils/locationUtils.js
@@ -12,3 +12,15 @@ export function getLocationZip(locationString) {
     return [parseInt(locationString)];
   }
 }
+
+export function zipcodeConversion(zipcode) {
+  try {
+    let convertedZipcode = zipcodes.lookup(zipcode);
+    let error =
+      'Sorry, zipcode is invalid. Please check your input and try again.';
+    if (!convertedZipcode) throw error;
+    return convertedZipcode;
+  } catch (error) {
+    console.log(error);
+  }
+}
